backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
gunicorn>=22.0.0
pydantic>=2.7.4
pydantic[email]>=2.0.0
groq>=0.9.0
python-dotenv>=1.0.1
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
firebase-admin>=6.0.0
supabase>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0
slowapi>=0.1.9
